+++ test/01-migrator/create/withSequentialMigrationNumberFormat.js
@@ -13,7 +13,13 @@ describe('migrator create with sequential migration number format', () => {
 		return Promise.resolve()
 			.then(() => {
 				return testUtils.createEnv({
-					migratorParams: {init: true, connect: true}
+					migratorParams: {
+						init: true,
+						connect: true,
+						configureParams: {
+							migrationNumberFormat: 'sequentialNumber'
+						}
+					}
 				});
 			})
 			.then((createdTestEnv) => {