+++ testUtils/createMigrator.js
@@ -1,3 +1,4 @@
+const _ = require('underscore');
 const Migrator = require('../lib/migrator');
 const removeMigratorDir = require('./removeMigratorDir');
 
@@ -10,7 +11,9 @@ module.exports = ({
 	const migrator = new Migrator();
 
 	return Promise.resolve()
-		.then(() => migrator.configure(configureParams))
+		.then(() => migrator.configure(_({
+			migrationNumberFormat: 'sequentialNumber'
+		}).extend(configureParams)))
 		// always connect migrator coz `removeDirBefore` makes unmark executed
 		.then(() => migrator.connect())
 		.then(() => {